pyproject.toml
[tool.poetry]
name = "bloom"
version = "0.1.0"
description = ""
authors = ["corgi-in-tights <59304120+corgi-in-tights@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
groq = "^0.11.0"
aiohttp = "^3.10.10"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
greenlet = "^3.1.1"
asyncpg = "^0.30.0"


[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
